Move IIRBuilder and AccessComputation to IIR directory
