+++ dawn/src/dawn/IIR/AccessComputation.cpp
@@ -12,7 +12,7 @@
 //
 //===------------------------------------------------------------------------------------------===//
 
-#include "dawn/Optimizer/AccessComputation.h"
+#include "dawn/IIR/AccessComputation.h"
 #include "dawn/IIR/AST.h"
 #include "dawn/IIR/ASTExpr.h"
 #include "dawn/IIR/ASTStmt.h"
