+++ dawn/test/unit-test/dawn-c/TestCompiler.cpp
@@ -17,11 +17,11 @@
 #include "dawn/CodeGen/CXXNaive-ico/CXXNaiveCodeGen.h"
 #include "dawn/CodeGen/CXXNaive/CXXNaiveCodeGen.h"
 #include "dawn/CodeGen/CodeGen.h"
+#include "dawn/IIR/IIRBuilder.h"
 #include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/SIR/SIR.h"
 #include "dawn/Serialization/SIRSerializer.h"
 #include "dawn/Support/DiagnosticsEngine.h"
-#include "dawn/Unittest/IIRBuilder.h"
 
 #include <gtest/gtest.h>
